requirements-pip.txt
# douban: pip install -i https://pypi.doubanio.com/simple/ --trusted-host pypi.doubanio.com <package>
django-model-utils
djangorestframework
django-cors-headers
django-guardian
django-rest-auth
django-filter>=2.0.0
django-mptt
easy_thumbnails
django-filer
diskcache
hendrix
aliyun-python-sdk-core
django-extensions